Added and updated unit tests.
