+++ cmake/Modules/add_polyamri_executable.cmake
@@ -1,7 +1,6 @@
 # This function adds a (serial) unit test executable to be built using cmockery.
 function(add_polyamri_executable exe)
   add_executable(${exe} ${ARGN})
-  set_target_properties(${exe} PROPERTIES OUTPUT_NAME ${exe} LINKER_LANGUAGE CXX)
-  target_link_libraries(${exe} ${POLYAMRI_LIBS})
+  target_link_libraries(${exe} ${POLYAMRI_LIBRARIES})
 endfunction(add_polyamri_executable)
