+++ cmake/Modules/add_polyamri_test.cmake
@@ -1,7 +1,7 @@
 # This function adds a (serial) unit test executable to be built using cmockery.
 function(add_polyamri_test exe)
   add_executable(${exe} ${ARGN})
-  target_link_libraries(${exe} cmockery ${POLYAMRI_LIBS})
+  target_link_libraries(${exe} cmockery ${POLYAMRI_LIBRARIES})
   set_target_properties(${exe} PROPERTIES COMPILE_FLAGS "-DCMAKE_CURRENT_SOURCE_DIR=\\\"${CMAKE_CURRENT_SOURCE_DIR}\\\"")
   add_test(${exe} ${exe})
 endfunction()
@@ -11,7 +11,7 @@ endfunction()
 # 1 test run will be generated for each processor number value.
 function(add_mpi_polyamri_test exe procs)
   add_executable(${exe} ${ARGN})
-  target_link_libraries(${exe} ${POLYAMRI_LIBS})
+  target_link_libraries(${exe} ${POLYAMRI_LIBRARIES})
   foreach (proc ${procs})
     add_test(${exe}_${proc}_proc ${MPIEXEC} ${MPIEXEC_NUMPROC_FLAG} ${proc} ${MPIEXEC_PREFLAGS} ${CMAKE_CURRENT_BINARY_DIR}/${exe} ${MPIEXEC_POSTFLAGS})
   endforeach()
