+++ polyamri/tests/CMakeLists.txt
@@ -1,7 +1,7 @@
 include(add_polyamri_test)
 
-#add_polyamri_test(test_tile test_tile.c)
-#add_polyamri_test(test_tile_set test_tile_set.c)
-#add_polyamri_test(test_grid_level test_grid_level.c)
-#add_polyamri_test(test_grid_hierarchy test_grid_hierarchy.c)
+add_polyamri_test(test_amr_patch test_amr_patch.c)
+add_polyamri_test(test_amr_patch_set test_amr_patch_set.c)
+add_polyamri_test(test_amr_grid_level test_amr_grid_level.c)
+add_polyamri_test(test_amr_grid_hierarchy test_amr_grid_hierarchy.c)
